Switch : Fix bug handling unknown output plugs
This is necessary to allow derived classes to inherit from Switch and add their own plugs. The primary problem was that `hash()` computed `inputIndex()` before calling `oppositePlug()`, which assumes that `oppositePlug()` isn't going to return null anyway. If the plug being hashed was an input to `indexPlug()`, this lead to infinite loops. We now pass in a context, and compute the input index lazily only when it is needed.