Dispatcher/DispatcherUI : Rationalise SubGraph support
The UI was checking for top-level TaskPlugs, whereas the Dispatcher was merely checking for Plugs, but considering nested plugs. The Dispatcher was checking that the plugs were actually connected to a TaskNode, but the UI wasn't bothering. Looking ahead, both are wrong : TaskNode-ness is irrelevant because TaskPlug provides the public API for task execution, and we should be passing TaskPlugs directly to the Dispatcher in the first place. That's a refactoring for another day though.

I'm really just updating this to demonstrate how the new range iteration is more convenient, and allows the same patterns to be used in both Python and C++ code. Over time, in idle moments, I'll migrate more of the existing codebase.