Serialiser : Use integer keys for ArrayPlug and RowsPlug children
This makes serialisations into better examples for how to use the API, and goes part way towards addressing #3575.

It also fixes a nasty serialisation bug with the Spreadsheet, as demonstrated by `SpreadsheetTest.testDeleteRowsAndSerialise()`. Deleting a row from the middle of the spreadsheet gives a break in the naming of the children - `[ "default", "row2" "row3" ]` in this case. But then when `addRows()` is used to recreate the spreadsheet on loading, it naturally creates a contiguous set of names - `[ "default", "row1", "row2" ]. This broke loading when we serialised using names, but is OK when serialising using indices.

An alternative fix to the Spreadsheet problem would have been to pass all the old row names to `addRows()`, but that's not really an API we want to expose if our goal is to move away from names anyway.

Fixes
-----

- Spreadsheet : Fixed bug that caused a load failure if a row had been deleted prior to saving.