ArrayPlug : Add resizeWhenInputsChange constructor argument
And `next()` method to return an unconnected child at the end of the array. The "automatic resize" behaviour of the ArrayPlug is really just a legacy from the days when we didn't have PlugAdders, and like all things that modify the graph in response to modifications to the graph, it's a little prone to unexpected consequences. This gives us the option of opting out of the resize behaviour, and eventually perhaps removing it entirely.

Breaking Changes :

- ArrayPlug : Added argument to constructor and additional private member data. ABI change only - source compatibility is maintained.