CompoundEditor : Don't save pinning using basic state
If an editor is pinned to specific node(s) when a layout is saved, we
used to store this as a simple boolean flag. When the layout was loaded,
we would then pin editors with this flag to nothing.

This seems to have been to allow layouts to restore with a viewer
'pinned off' to help with large scenes. As we now have async updates, as
well as a pause button, this behaviour is generally unhelpful.

There is a bigger question as to the potential benefit of saving
layouts with scripts (where storing a specific node reference could make
sense). Coupled with the numerous recent changes to the Editor Focus
mechanism (that allow saved layouts to express some specific pinning
states when loaded). It feels this simple pinning restoration is no
longer needed.

This commit removes support for the old `pinned` flag. Existing layouts
containing this flag will still load and the flag be ignored.

Improvements
------------

- Layouts : Prevented editors pinned to a specific node from restoring
  with the editor pinned to an empty node set. Specific pinning is no
  longer saved or recalled.