CompoundEditor : Store editor node set driver in layouts
Allows a NodeSetEditor's driver to be saved in a layout.

We achieve this by deriving a layout-specific 'path' to an editor based on
its nesting, which we then use to reference specific editor instances when
loading the layout.

Though this leaks the widget hierarchy into the layout serialisation, we
are already doing this in the way we store the child editors anyway.

In the longer-term, we're going to most likely migrate Editors over to
GraphComponents, so this intermediate implementation endeavours to keep
all the save/restore logic contained within CompoundEditor.