CompoundEditor : Preserve editor links when dropping nodes on an editor
When editors are linked, it is more intuitive for users that a node drop
updates the driving editor, rather than breaking the link. Users generally
make use of node drop to update what is seen in an editor and this is a
more common intention than wanting to change the link-state of say a
viewer and hierarchy view.

User facing changes :

 - Dropping a node onto an editor no longer breaks editor links and
   instead updates the driving editor.