Remove Travis CI
Azure is now our standard CI platform, and provides more complete testing than Travis (GafferArnold is included, we test on a Docker image aligned with VFXPlatform etc). The only real downside here is that we're dropping testing with GCC 4.8, which was dropped by VFXPlatform in 2018, but lives on in various DCCs. If anyone wants to maintain a Docker image and set of dependencies for this platform, we'll happily add it to the testing matrix, but we can't justify doing the work ourselves any more.