CatalogueUI : Ensure new images are added at the end of the catalogue
Previously, the catalogue UI would insert new images at index 0. This
resulted in images added whilst the UI was closed appearing near the
beginning, rather than at the end.

This commit adjusts the sort logic such that images without any specific
metadata are ordered as to their plug position. We also no longer add
metadata when added in the UI so that the state is consistent whether the
UI is open or closed.