Capsule : Guard against concurrent access to signals
As documented in the comments, this is not watertight, but it should fix the bug reports that we have in practice. It's still possible to get a crash if the UI accesses the same signal being accessed concurrently by a capsule, but the chances of this are small compared to the chances of two capsules accessing at the same time. Of the alternative fixes suggested in the comments, I'm rather drawn to the third one.