GafferScene : Add ObjectProcessor base class
This provides a simpler means of processing objects than SceneElementProcessor, with the following benefits :

- More obvious what it's for.
- Just 3 virtual functions to be overridden, rather than a potential 12 for SceneElementProcessor.
- It's possible to pass through the input hash without paying the cost of initialising with a default hash first.
- Follows Gaffer's standard hashing pattern of calling the base class first, rather than the non-standard pattern
  in SceneElementProcessor.
- `affectsProcessedObject()` makes it easier to keep dirty propagation syncronised with hash and compute.
- Not filtered to anything by default. SceneElementProcessor is the opposite, we don't like it, and we have to
  remember not to use the default when writing any new nodes.

I did also consider removing the `inputObject` from the signature for `computeProcessedObject()`, and omitting its hash from `ObjectProcessor::hashProcessedObject()`. This would allow DeleteObject (and a future CopyObject) to derive from ObjectProcessor without paying for unwanted input evaluations. In the end, I decided that it was better to keep things simpler for the vast majority of derived classes, and have DeleteObject continue to be a special case of its own.