Viewer : Allow up/down arrow keys to navigate the catalogue
Ideally we'd just be updating catalogue["imageIndex"], but this wouldn't
take into account the custom ordering supported by the CatalogueUI. Long
term we need to make the public catalogue API (via its plugs) match the UI,
but its out of scope for this. For now at least, we keep the magic in one
confined place.

User facing changes:

 - When the mouse is over the Viewer, and the Viewer is showing a Catalogue
   (or any other node downstream of one), the up and down arrow keys can
   be used to navigate images in the catalogue, as they do when in the
   catalogue's Node Editor.