CompoundEditor : Handle out-of-order deletion of python widget wrappers
Switching layouts with linked editors could cause 'c++ object already
deleted' errors when an outgoing driving editor was deleted before
its slaves.

We have to inform slave editors they are no longer driven so they can
update their UI. In this case however, thanks to #3179, the UI that
needs to be updated has already been deleted on the c++ side - so updates
throw.

As there is no way for the python objects to know of their impending
deletion, this last-ditch sanity check seems the only valid way to
handle this situation.