Parent : Accept multiple child inputs
The approach taken here is to use the new ChildNamesMap to perform the necessary merging and uniquefying of child names. This is around 30% faster than an initial prototype that just used an internal Group/Subtree combo to do the merge, and around 12% faster than using a hypothetical MergeRoots node.

Fixes #1202.