NodeSetEditor : Ensure pinned state is always reflected in the title
Previously, if an editor's node set had no applicable nodes, the
square-bracket ` [<nodes>]` suffix would be omitted.

This made it look like the editor wasn't pinned. Which was a lie, and was
particularly noticable when in the default layout, the Viewer was pinned
to a 2D node, so the 3D editors had no applicable nodes in their set.

Now the title will always contain the bracket suffix, even if there are
no nodes. In the case of no nodes, it will simply be ` []`.