TaskMutex : Remove workaround for TBB prior to 2018 Update 3
The downside of this is that not all worker threads might be used by `TaskMutex::execute()`. But the upside of this is that we're no longer exposed to horrendous slowdowns in TBB if we were to use `TaskMutex::execute()` frequently, and `TaskMutex::execute( f )` is guaranteed to call `f` on the current thread, removing the need for nasty workarounds in ValuePlug.