IE Build Options : Use the Gaffer registry for standalone builds.
I'm not sure why this was ever set to cortexReg, but it seems the only value coming out of that was tbbVersion, which already has a fallback to the cortexReg if its missing from the targetAppReg.

Note I left the pythonVersion coming from the Cortex Registry as we can't have those differ (yet).