PlugValueWidget : Don't allow creation of inputs to output plugs
An input connection would mean that the plug value is no longer computed as expected, so we now disallow its creation. There might be an argument that we should allow connections to be made in the case of an output plug on a Box, but since the only way to create an output plug is via promotion, and promotion sets the input anyway, I don't think this is necessary.