Handle : Support holding `<shift>` for precise motion drags
Any class that makes use of `Handle::LinearDrag` and `Handle::PlanarDrag`
gain support for 'precise' mode. When the `shift` key is held down,
subsequent mouse movements results in 10x smaller changes in the resulting
position.

Note: Scaling is applied in 'position' space, rather than in mouse space.
This generally feels to work predictably for the simple cases of linear
and planar drags. We may find use cases where this doesn't work and we
instead need to just do this right in mouse space.

Implements part of #3324.