Build : Ensure ARNOLD_LOGIN and ARNOLD_PASSWORD are set
Aborts Arnold install if ARNOLD_LOGIN or ARNOLD_PASSWORD aren't set,
instead of spamming the Autodesk servers.

It seems difficult to get Azure to not set it. So we also handle it being
a string starting with a $ too. It does mean you can't have a password that
begins with a dollar, but lets cross that bridge...