TabbedContainer : Fix exception when last tab is removed
In a previous commit, we fixed the order of change signals and widget
removal (such that the change signal would be given the correct widget).
Now we have this the right way around, we need to make sure that when the
last widget has been removed (and the change signal is emitted) we handle
the container being empty.

NOTE: Listeners to `currentChangedSignal` need to be able to handle the
supplied widget being `None` (which has always been the case, just it wasn't
exercised in practice previously).