GraphEditor : Track changes to parents of the current root
The validity of the current root, and the title of the editor depend upon
not only the root itself, but its hierarchy back up to the script node.

We need to make sure we track all ancestors for changes, otherwise we can
easily end up in a root that has been removed from a script without
realising.

Fixes #3455.