StandardLightVisualiser : Simplify color indicator
Simplifies the light color indication to be a smaller, camera-facing
dot. User feedback suggested that the exposure indicator was often too
small on screen to be useful. The majority of users were in favour of
less clutter in the viewport.

In addition, we now only color the indicator with the light's tint, and
don't factor in intensity or exposure. In most real-world cases, this
resulted in either bleached out, or too dark a display. It also now
matches texture rendering for area lights.

Improvements
------------

 - Viewer :
     - Simplified light the light color indicator and removed the
       exposure wedges.
     - Removed the contribution of intensity and exposure to the color
       indicator.