RendererController : Defer updateObject hash update
We were inadvertently updating the hash before successfully updating the
object. If this took some time, and was cancelled, then the hash would
match the next time the update ran and we'd be stuck with an empty
object representation.

This started to manifest itself whilst implementing #3407, as the shading
time to generate light textures overlapped with cancellations due to calls
to `GafferSceneUI::ContextAlgo::setExpandedPaths`.

Fixes #3469.