Appleseed : Allow nodes to be hidden from the UI
Removes Appleseed nodes from the node menu as well as output presets and
diagnostic shading modes when `GAFFERAPPLESEED_HIDE_UI` is set to `1`.

For facilities not using Appleseed in production. There are a lot of
name collisions with Arnold nodes that makes finding the right one harder.

We still load Appleseed so it can be used for OSL shader previews, etc.