Set : Deprecate `paths` plug
The original version of the Set node didn't have a `filter` plug,
so the only way of specifying locations was to enter them as paths.
This had very good performance, but allowed users to accidentally
enter invalid locations, which could have unintended consequences
downstream.

Later on, we added the `filter` plug to provide a more flexible way
of specifying locations. This method guarantees that only valid locations
are used, but came with an overhead in performance. We documented that
users should prefer the `paths` plug where possible.

In practice, users are just confused about which option to use and when.
We've also substantially improved the performance of the `filter` option
since its introduction, to the point where there is negligible overhead
in replacing `paths` plug usage with `PathFilter` usage. So, we wish to
rip out the `paths` plug to remove confusion and guarantee that sets can
never contain invalid paths.

Here we're merely deprecating the `paths` plug and hiding it from the UI
unless it is already in use (has a non-default value or current input).
In a later version we'll remove it fully, at which point we'll deal with
the problem of auto-converting old `paths` usage into PathFilters.

Improvements
------------

- Set : Hid unused `paths` plugs from the UI, in preparation for eventual
  removal of the plug itself. The `filter` plug should now always be used
  in preference to the `paths` plug.