GafferOSL : Add algo functions to aid rendering OSL networks to images
`ShadingEngineAlgo::shadeTexture` is deliberately made non-cancelable so
it can be used in code that is not cancel safe (ie: GL rendering).

Long-term, we plan to revamp ShadingEngine and its API such that it doesn't
make direct use of the current context, and instead requires any globals
or cancellers to be passed in through the API explicitly.

In the short term, we ensure cancellation can't occur by scoping a default
context in `shadeTexture`. This does mean that for now, shaders won't have
access to any context data (including the current frame). It does however
allow us to employ this in the Viewer, where cancellations are common.