OpenGLRenderer : Classify visualisation renderables
In many situations, visualised locations have a non-uniform scale in
their transform. This is common with area lights for example where the
transform determines the lights proportions. Visualisers have no access
to the transfrom stack so are unable to compensate for this internally.

This commit introduces a `VisualiserType` that can be used to categorise
the renderables returned by a visualiser. Two types are defined:

 - Geometry : Renderables will inherit the locations full transform.
 - Ornament : Renderables will only inherit the locations translation.

API
---

- IECoreGLPreview : Added `VisualisationType` and `VisualisationMap` to
  allow classification of visualisation renderables returned by visualisers.

Breaking Changes
----------------

- IECoreGLPreview : Refactored the visualisation methods of
  `LightVisualiser`, `LightFilterVisualiser` and `AttributeVisualiser`
  to support categorisation of renderables via `VisualisationMap`.