SetAlgo : Improve treatment of wildcards in object names
Prior to this, `/*/sphere` would parse as `( / | * ) | /sphere`, which is almost certainly not what was intended. This was an unwanted side effect of telling the parser that object names shouldn't contain wildcards; it forced the parser into parsing the middle of the object name as a set name.

This commit actually makes the parser dumber by removing the distinction between set names and object names, so that it just parses "identifiers". We can then distinguish between objects and sets ourselves in `AstHasher/AstEvaluator`. This allows us to emit a more descriptive error message if an object name contains wildcards.