Viewer : Improve clipping plane fitting
In production, we were still seeing many cases where the near clip plane
would be set too far from the camera, such that dollying in (or using 'f'
to frame selection) would clip the scene, or achieve bad framing.

This moves to a simpler approach, whereby we keep the near plane much
closer to the camera.

Fixes
-----

- Viewer : Improved clipping plane "Fit ..." command logic to avoid
  clipping the scene when dollying in, or bad framing using <kbd>f</kbd>.