StandardLightVisualiser : Visualise area light spread
Arnold implements spread as a 0-1 angular range rather than anything
in light space, as such we need to use the ornament visualisation type
so it isn't distorted by light scaling. This would mis-represent the
rendered result. Consequently, we're limited to a relatively naive
presentation, but it gives a meaningful representation of the final
appearance.

Improvements
------------

 - Viewer : Added an approximation of Arnold area light spread.