CatalogueUI : Ensure single source of truth for image ordering
Moves to a single point for ordering images (_ImagesPath). Hopefully
this will result in consistent behaviour, even when there are collisions
with stale metadata and/or plug indices.

In order to do this, it also moves some parts away from querying
metadata directly. This may result in a less efficient implementation,
but hopefully a more stable one.