CatalogueUI : Insert duplicated images before original
Duplicate is usually used to snapshot in progress renders. As such, it's
helpful to place the duplicates before the source, rather than at the
end of the list - as the source will soon be more up-to-date than the
copy. People often refer to it as a 'snapshot' button, rather than
duplicate.

Improvements
------------

- Catalogue : Changed image duplication behaviour so the copy is now
  placed above the source in the image list. If <kbd>alt</kbd> is held
  whilst duplicating, then Catalog will switch to the last duplicate.