TransformTool::Selection : Make all members private
This is just renaming things and shuffling code around a bit - there should be no functional changes. The motivation is that EditScopes are going to require Selection objects to have more complex state, which we intentionally mutate. Disallowing unintentional mutations and allowing state checks on access will improve clarity at that point.