Plug popup menus : Connect with `scoped = False`
This represents a _tiny_ bit of progress towards _maybe_ one day being able to reverse the default for `scoped` to match the C++ behaviour. That's a rather ambitious goal because we'd need all Gaffer client code to be updated to specify a `scoped` argument before we can do it. But the fewer examples we provide of how _not_ to do it, the better.