SceneAlgo : Include intermediate connections in `history()`
This is what we need to be able to use `history()` to make a more accurate SceneInspector one day. It also has the nice property that we never return null.

I did consider adding an extra field to History, to denote whether or not this part of the history is a connection or a compute. I decided against, because that is all queryable direct from the horses mouth via the plug (if it's an unconnected output, it's a compute). We may want to add a `PlugAlgo::isComputed( plug )` utility method to help with this at some point, as we have several other use cases for such a query.

API
---

- SceneAlgo : `history()` now returns History items for all upstream plugs in the history chain, not just those where a computation was performed.