CatalogueUI : Prefer NodeAlgo to homegrown graph traversals
This homegrown traversal had the common problem that it didn't keep track of nodes already visited. This can lead to excessive search time for certain graph layouts.

This is the second time I've made this commit. The first time was thwarted by the fact that the first implementation of `findUpstream()` didn't traverse inside Boxes (#3605).