EventLoop : Fix application style
The attempt to choose the "plastique" style was failing, and presumably has been ever since Qt 5. We now use the "Fusion" style to give us a consistent baseline regardless of platform, and that fixes problems such as the centre aligned tab bar on Mac reported in https://github.com/GafferHQ/gaffer/pull/3566. We now also assert that setting the style has succeeded, so we won't be caught out again in the same way.