SpreadsheetUI : Fix bugs in row width handling
We were assuming that we'd always find a "spreadsheet:rowNameWidth" metadata entry because it was registered on the Spreadsheet base class, but if a RowsPlug was added to another node it would be missing. We were failing to retrieve the current width properly in the row width menu because we were querying it from the wrong row (it is stored on the default row). Moving all metadata access into helper methods makes it easier to get this right.

Fixes
-----

- Spreadsheet :
  - Fixed bug that could break promoted Spreadsheet UIs.
  - Fixed bug that meant the current width was not checked in the row name width popup menu.