IECoreGLPreview : Refactor visualisation mechanism
Refactors Visualisers such that they return a vector of `Visualisation`
structs instead of a type-keyed array.

In practice, as we developed more visualisers and attempted to match
behaviour across renderers, it became clear that the concrete-type
approach wouldn't scale to all the permutations required. Particularly
as specific rendering behaviours don't match to specific user-land
categories such as 'frustums'. The requirement to expose drawing
toggles in the UI made the typed approach untenable.

This commit moves to returning visualisation objects, that define
scaling and framing behaviour, and a category - leaving it up to the
renderer to filter these accordingly. User-land toggles then act on
specific categories.

A subsequent commit will conform the use of the new behaviour options
in existing visualisers.

API
---

- IECoreGLPreview : Added support for more flexible visualisations via
  the `Visualisation` struct (#3569).

Breaking Changes
----------------

 - IECoreGLPreview : Changed the return type of `(Light|LightFilter|Attribute)Visualiser`
   classes. Visualisations are now a vector of `Visualisation` structs
   (#3569).