ImageTransform : Fix inconsistent filtering of `rotate!=0 && translate==0`
The ImageTransform node separates out the translation and scaling components of the transform from the rotation component, and applies them using an internal Resample node. This provides improved performance for changes in scale by taking advantage of separable filtering. The ImageTransform also deliberately performs this resampling _even for identity transforms_, to avoid sudden jumps in situations such as an animated translation passing through the origin on one frame, or scaling passing through a value of 1.

This consistency of filtering was being broken in the case of non-zero rotation without translation. In this case the resampling was bypassed entirely, leading to jumps in filtering when a rotation was initiated. To fix this we now _always_ use the resampled input for all cases. This does lead to slower operation for pure-rotation operations, but that is consistent with the hit we've already accepted for the identity case.

Fixes
-----

- ImageTransform : Fixed inconsistent filtering of transforms containing rotation but no translation.