ImageView : Add CropWindowTool and improve message presentation
A note on presentation changes:

Using the overlay box is more consistent with other tools and works
better when the view can be panned/zoomed as per ImageView.

It caused a problem with the ImageView though, as the color inspector
is center-aligned, but the crop overlay is left. As we don't want to
have to adjust the CropWindowToolUI presentation based on which view
is hosting, we compromise by making the color inspector wider, so it
aligns on the left side, but keeping the color readouts centred so they
are closer to the view point.

Ideally we'd have view safe areas so we don't need to do this all with
hard-coded margins, etc.

Had some issues using the spacer-widget trick employed by other tools, a
more robust solution seemed to be to instead make use of stylesheet
margins to create the appropriate inset.  We can't use these right now
for Transform tool as there are other plug value widgets in there that
aren't uniquely accessible via a suitable selector.

Features
--------

 - Viewer :
   - Added the Crop Window Tool to image views to allow a scenes crop
     window to be adjusted directly from the rendered image (#2835).
   - Changed Crop Window Tool status presentation to match other
     tools (#2835).