InteractiveArnoldRenderTest : Relax equality test for Arnold 6
Sampling changes appear to prevent convergence to 1 for all pixels with the low sampling settings used in this test. Allowing a small delta has no effect on the accuracy of this test, because what we are really asserting is that the light is linked _at all_, and failure would be indicated by a completely black image.

Also removed unnecessary node. This had no impact on the outcome of the test.