GafferSceneUITest : Add viewport rendering performance test cases
The singular goal of these tests is to rest the impact of changes that
may affect the performance of the renderer's `render()` method, such as:

 - Changes to ObjectVisualisers
 - Changes to LightVisualiser or LightFitlerVisualisers
 - Changes to IECoreGLPreview/Renderer itself

It deliberately excludes scene generation mechanics or widget nesting
considerations that may affect the user experienced performance in
real application usage.

These tests are generally not run under CI, where the performance of
the hosting hardware varies too greatly to determine a reasonable high
water mark. The do however report to `Gaffer.TestRunner.PerformanceScope`
to allow comparison to historical runs.

API
---

- GafferSceneUITest : Adds `SceneUITestCase` to aid render performance
  test authoring (#3606).