Revert "CatalogueUI : Prefer NodeAlgo to homegrown graph traversals"
This reverts commit 1de95ebe54ac2f5ef4675d3e8b4a23a342a79cd3.

Fixes #3605. The problem is that `upstreamNodes()` doesn't traverse inside Boxes, whereas our homegrown traversal did. We either need to introduce an `upstreamPlugs()` family of methods mirroring `upstreamNodes()` but at the plug level, or we need to introduce some control over  traversal of internal nodes to `upstreamNodes()`. While we mull that over, we're simply reverting the problematic commit.