Viewer : Left-align ColorInspector in ImageViews
As part of #3582 we stretched out the `ColorInspector` so the info
could still remain centred, whist the `CropWindowTool` was suck on the
left due to no view-dependent placement.

The more I used it, I wasn't entirely sure about the huge margins.

This left-aligns it with the other tool overlay.

Long term, we want to implement safe-margins in the viewer so we can
manage all this a little more intelligently.

Improvements
------------

- Viewer : Left-aligned the image ColorInspector to remove huge margins.
  [since 0.56.0.0b2]