MenuBar : Allow subMenus to opt out of short cut discovery
Facilities often register custom menu bar menus that talk to asset
management systems, etc. that can cause (relatively) slow build times.

Qt's ShortcutOverride event bubbles up the widget hierarchy before
KeyPress events are sent to the target widget. We use this in
`GafferUI.MenuBar` to implement global short cuts. To achieve this
we fully build the main menu bar and search for any actions with
short cut keys.

This effectively means we build the main menu bar whenever the user
presses a key over an editor without a text input focus. A common case
for this is `tab` to open the Node Menu, which can result in long
delays before the menu appears.

We can't avoid this behaviour (as we want main menu items with short
cuts), so we instead allow expensive menus to opt-out of short cut
discovery if they don't have any.

Closes #3651.

API
---

 - MenuBar : Added support for `hasShortCut` on a `IECore.MenuDefinition`
   which when `False` prevents a dynamic sub menu from being build to
   discover keyboard short cuts (#3651).