MergeScenes : Optimise `computeActiveInputs()`
This knocks 60% off a benchmark that merges two sparsely overlapping scenes.

Note that this also fixes a bug in `hashActiveInputs()`. On one code path we were calling `append( connectedInputs() )` and on another `append( indexOfAnActiveInput )`. In the latter case, if there was a single active input, we could append the exact same bit pattern as the `connectedInputs()` case and end up with a hash conflict.