Range/RecursiveRange : Use `children()` rather than `__getitem__`
This protects against a problem whereby `startup/GafferScene/copyAttributesCompatibility.py` intentionally returns `self` for `CopyAttributes.in[0]`, leading to recursion. Arguably we should fix the compatibility config so it only does its magic while a file is loading, but we have no way of knowing if people are relying on its current behaviour in their own code.

Fixes
-----

- Range/RecursiveRange : Fixed problems caused by visiting CopyAttributes nodes.