NameWidget : Support `readOnly` metadata
This fixes a bug whereby the editability of the node name in the NodeEditor didn't update when a node was locked. Note that to make this possible, we need to disable the propagation of `NodeEditor.setReadOnly()` to the NameWidget, so we are effectively trading one bug for another. Fortunately the latter bug is irrelevant because it seems we never call `NodeEditor.setReadOnly()`. In fact, we should probably remove that method one day, along with the also-unused `PlugValueWidget.setReadOnly()`.

Fixes
-----

- NodeEditor : Fixed bug that allowed the node name to be edited on a locked node.