Plug/Node Widgets : Remove `set/getReadOnly()` methods
These are not used at all in the Gaffer codebase, so represent an unnecessary burden when subclassing or extending, particularly for custom PlugValueWidgets. I am about to make the PlugValueWidget API a little more complex for subclassing, and removing `set/getReadOnly()` will help to keep the overall burden lower.

One of the most burdensome aspects of this was the necessity of mirroring a parent widget's state onto all of its children. If we were ever to reintroduce a mechanism like this, we'd want it to be inherited through the widget layout hierarchy automatically somehow.