CatalogueUI : Make columns configurable
As part of the implementation of #3646, we need to make the catalogue
columns configurable. This commit switches the current, single
'name' column over to the registry mechanism, adding the CatalogUI
column API.

The goal is to allow any column to be extended or replaced as needed
in a facility, but to provide a sensible starting point in stock
Gaffer.