ArnoldTextureBake : Simpler task allocation
Instead of considering the number of objects that need baking, it just allocates tasks to UDIMs.  This could be a worse metric in certain corner cases where there are vast numbers of objects in one UDIM, but in general, UDIMs with more objects have objects that cover fewer pixels, so it works out OK.  This simpler metric gets rid of some weird results where tasks would be allocated badly in practical examples.