TransformTool : Improve status display
The Selection class already provided a lot of useful warnings but they were lost because we were clearing the selection if anything non-editable was found. We now keep the full selection so we can display the warnings, and have fixed the spots where code was assuming all non-empty selections were editable.