PathListingWidget : Allow sortability to be expressed in constructor
On some platforms, not doing this and calling `setSortable( False )` after
construction could result in the constructor supplied path already being
sorted - see #3684.

As the concept of sortablility in this widget is a little flawed (it
interacts with the underlying model), and adjusting the sortability on the
fly doesn't make a lot of sense (as you can't easily revert to the original
state). As such, this commit also deprecates `setSortable` and `getSortable`.