PlugLayout : Build immediately on construction
This allows us to avoid a bunch of hacks where we were getting really poor initial sizes for windows. In practice the lazy building wasn't doing much for us, as our most expensive UIs (things like the parameters for aiStandardSurface) are all in a single layout that is immediately visible on editing the node.