PathFilter : Require `scene:path` in `computeMatch()`
We added a special case for missing `scene:path` back in 7627785d08fed66d317a7f0f5433b5ea3774653b, because serialisation was triggering computes without a suitable context. This was the wrong fix though. The right fix, which we made subsequently, was to avoid triggering any computes during serialisation. As a result, `PathFilterTest.testCopyPaste()` passes without needing the special case in PathFilter.

Note that we still need the special case for missing paths in `computeHash()`, but that is because we use it as a way of requesting a "whole scene hash" in Isolate and Prune. It would be good to find a way of tidying that up too, but it's a different problem.