SpreadsheetUI : Fix row name sizing bug
This was introduced by 3a8ba8b83c3a1d1ad24405ba8a5448785cb789e7, with the root cause being an attempt to apply the column widths for the main columns to the two columns in the row name section. Added similar safeguards to the application of section order and column visibility.