PathFilter : Add `roots` filter plug
This can be used to specify multiple scene locations that the `paths` are specified relative to.

I considered a couple of different approaches to the one taken in the end :

1. Using FilterResults internally to compute all the roots from `rootsPlug()`. This may have resulted in a simpler implementation, but it has the downside of triggering a potentially large computation across the whole scene, just to compute one filter result. This cost would have been amortised across later filter computations, but only if they were actually requested. To protect the FilterResults from excessive hashing, we would have needed to construct a clean context, which is equivalent to the main cost in the current implementation anyway.
2. Using a brute-force approach that queried every ancestor location for matches against `rootsPlug()`, every time we computed a filter result. This would have scaled poorly for deep hierarchies.

The current approach is an optimisation of 2) where we cache results from ancestors on `rootSizesPlug()` so that at any given location we only have the overhead of one additional compute.