Catalogue : Fixed invalid name handling in `Image::load()`
If we don't generate a valid name, we get an exception when the user either loads an image from disk or drags an image node into the Catalogue. Worse still, because Python holds on to the last exception, this could keep the UI in a zombie state after the Qt half was deleted, leading to later crashes.

Really we should be able to address the sanitisation of names more centrally in GraphComponent, but that is a task for another day. The original restrictions were based on a need for the names to be valid Python identifiers, so we could access plugs as `NodeName.plugName`, but since we now use `NodeName["plugName"]` we could be much more generous. We should really support UTF8 encoded strings as well, the main obstacle to dealing with that being the rendering of text in OpenGL.