SpreadsheetUI : Avoid circular references
This is a slightly speculative fix for a series of errors I discovered in my shell after testing #3560 :

```
Traceback (most recent call last):
  File "/disk1/john/dev/build/gaffer/python/Gaffer/WeakMethod.py", line 67, in __call__
    return m( *args, **kwArgs )
  File "/disk1/john/dev/build/gaffer/python/GafferUI/PlugLayout.py", line 516, in __plugDirtied
    if not self.visible() or plug.direction() != plug.Direction.In :
  File "/disk1/john/dev/build/gaffer/python/GafferUI/Widget.py", line 259, in visible
    return self.__qtWidget.isVisibleTo( relativeTo )
RuntimeError: Internal C++ object (PySide2.QtWidgets.QWidget) already deleted.
```

I was unable to get this to reproduce, but the error is a classic symptom of a circular reference keeping a `GafferUI.Widget` alive past the point the Qt half has expired. It seems more than coincidence that it occurred while testing new code that did indeed contain a circular reference due to baking the NodeEditor into a menu command.

The fix is pretty straightforward : just rely on the automatically provided `menu` argument to commands instead of stashing references inside a `partial()`.